requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
apscheduler>=3.10.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
alembic>=1.13.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
